Remove nginx from Docker config for frontend.
On most machines, quran.com does not live alone, sharing the hardware
with other services. Consequently, there is typically already an
instance of nginx running. Instead of having 3 instances of nginx
running on such machines (1 for the frontend, 1 for the api, and 1 for
whatever other services are running), this change allows us to merge the
frontend nginx with the non-api nginx config.