Fix node-http-proxy host in proxy request (#501)
node-http-proxy was putting a host header with the current server name,
which was thus causing nginx on the receiving end to send the request to
the default handler (or the first one). This patch asks node-http-proxy
to rewrite the host and port, and also to change the origin of the host
header to onequran's url.