Fixes #547 Support continuing from last visited ayah (#559)
* Fix last visited functionality
- Read last visited surah/ayah from cookie and show resume option on main page if cookie exists
- Refactor and cleanup previous last visited code to be more readable and match the new UI

* Update lastVisit cookie ayah requests, switch cookie ayah field to ayahId for consistency

* Delete last visit styles.scss

* Change resume heading to continue and support urdu/arabic translations

* Move vertical lines from class title into new class

The existing .title class made it difficult for titles that are not a series of items to use the span tag as it would add unwanted vertical lines. The only user of this was QuickSurahs component. I moved the vertical lines into a new "items" class so that all other users which are not a series of items can use the title class without getting vertical lines in spans. Only last visit benefits from this at the moment since the main Home component doesn't have any spans inside the title and there are no other users of this scss file.

* Add debug line back in lastvisit

* Minor refactoring